bring select columns to the front of a dataframe

bring select columns to the front of a dataframe

bring_to_front(.df, ...)

Arguments

...

columns to bring to front

df

dataframe

Examples

Theoph %>% bring_to_front(conc, Time)

#>      conc  Time Subject   Wt Dose
#> 1    0.74  0.00       1 79.6 4.02
#> 2    2.84  0.25       1 79.6 4.02
#> 3    6.57  0.57       1 79.6 4.02
#> 4   10.50  1.12       1 79.6 4.02
#> 5    9.66  2.02       1 79.6 4.02
#> 6    8.58  3.82       1 79.6 4.02
#> 7    8.36  5.10       1 79.6 4.02
#> 8    7.47  7.03       1 79.6 4.02
#> 9    6.89  9.05       1 79.6 4.02
#> 10   5.94 12.12       1 79.6 4.02
#> 11   3.28 24.37       1 79.6 4.02
#> 12   0.00  0.00       2 72.4 4.40
#> 13   1.72  0.27       2 72.4 4.40
#> 14   7.91  0.52       2 72.4 4.40
#> 15   8.31  1.00       2 72.4 4.40
#> 16   8.33  1.92       2 72.4 4.40
#> 17   6.85  3.50       2 72.4 4.40
#> 18   6.08  5.02       2 72.4 4.40
#> 19   5.40  7.03       2 72.4 4.40
#> 20   4.55  9.00       2 72.4 4.40
#> 21   3.01 12.00       2 72.4 4.40
#> 22   0.90 24.30       2 72.4 4.40
#> 23   0.00  0.00       3 70.5 4.53
#> 24   4.40  0.27       3 70.5 4.53
#> 25   6.90  0.58       3 70.5 4.53
#> 26   8.20  1.02       3 70.5 4.53
#> 27   7.80  2.02       3 70.5 4.53
#> 28   7.50  3.62       3 70.5 4.53
#> 29   6.20  5.08       3 70.5 4.53
#> 30   5.30  7.07       3 70.5 4.53
#> 31   4.90  9.00       3 70.5 4.53
#> 32   3.70 12.15       3 70.5 4.53
#> 33   1.05 24.17       3 70.5 4.53
#> 34   0.00  0.00       4 72.7 4.40
#> 35   1.89  0.35       4 72.7 4.40
#> 36   4.60  0.60       4 72.7 4.40
#> 37   8.60  1.07       4 72.7 4.40
#> 38   8.38  2.13       4 72.7 4.40
#> 39   7.54  3.50       4 72.7 4.40
#> 40   6.88  5.02       4 72.7 4.40
#> 41   5.78  7.02       4 72.7 4.40
#> 42   5.33  9.02       4 72.7 4.40
#> 43   4.19 11.98       4 72.7 4.40
#> 44   1.15 24.65       4 72.7 4.40
#> 45   0.00  0.00       5 54.6 5.86
#> 46   2.02  0.30       5 54.6 5.86
#> 47   5.63  0.52       5 54.6 5.86
#> 48  11.40  1.00       5 54.6 5.86
#> 49   9.33  2.02       5 54.6 5.86
#> 50   8.74  3.50       5 54.6 5.86
#> 51   7.56  5.02       5 54.6 5.86
#> 52   7.09  7.02       5 54.6 5.86
#> 53   5.90  9.10       5 54.6 5.86
#> 54   4.37 12.00       5 54.6 5.86
#> 55   1.57 24.35       5 54.6 5.86
#> 56   0.00  0.00       6 80.0 4.00
#> 57   1.29  0.27       6 80.0 4.00
#> 58   3.08  0.58       6 80.0 4.00
#> 59   6.44  1.15       6 80.0 4.00
#> 60   6.32  2.03       6 80.0 4.00
#> 61   5.53  3.57       6 80.0 4.00
#> 62   4.94  5.00       6 80.0 4.00
#> 63   4.02  7.00       6 80.0 4.00
#> 64   3.46  9.22       6 80.0 4.00
#> 65   2.78 12.10       6 80.0 4.00
#> 66   0.92 23.85       6 80.0 4.00
#> 67   0.15  0.00       7 64.6 4.95
#> 68   0.85  0.25       7 64.6 4.95
#> 69   2.35  0.50       7 64.6 4.95
#> 70   5.02  1.02       7 64.6 4.95
#> 71   6.58  2.02       7 64.6 4.95
#> 72   7.09  3.48       7 64.6 4.95
#> 73   6.66  5.00       7 64.6 4.95
#> 74   5.25  6.98       7 64.6 4.95
#> 75   4.39  9.00       7 64.6 4.95
#> 76   3.53 12.05       7 64.6 4.95
#> 77   1.15 24.22       7 64.6 4.95
#> 78   0.00  0.00       8 70.5 4.53
#> 79   3.05  0.25       8 70.5 4.53
#> 80   3.05  0.52       8 70.5 4.53
#> 81   7.31  0.98       8 70.5 4.53
#> 82   7.56  2.02       8 70.5 4.53
#> 83   6.59  3.53       8 70.5 4.53
#> 84   5.88  5.05       8 70.5 4.53
#> 85   4.73  7.15       8 70.5 4.53
#> 86   4.57  9.07       8 70.5 4.53
#> 87   3.00 12.10       8 70.5 4.53
#> 88   1.25 24.12       8 70.5 4.53
#> 89   0.00  0.00       9 86.4 3.10
#> 90   7.37  0.30       9 86.4 3.10
#> 91   9.03  0.63       9 86.4 3.10
#> 92   7.14  1.05       9 86.4 3.10
#> 93   6.33  2.02       9 86.4 3.10
#> 94   5.66  3.53       9 86.4 3.10
#> 95   5.67  5.02       9 86.4 3.10
#> 96   4.24  7.17       9 86.4 3.10
#> 97   4.11  8.80       9 86.4 3.10
#> 98   3.16 11.60       9 86.4 3.10
#> 99   1.12 24.43       9 86.4 3.10
#> 100  0.24  0.00      10 58.2 5.50
#> 101  2.89  0.37      10 58.2 5.50
#> 102  5.22  0.77      10 58.2 5.50
#> 103  6.41  1.02      10 58.2 5.50
#> 104  7.83  2.05      10 58.2 5.50
#> 105 10.21  3.55      10 58.2 5.50
#> 106  9.18  5.05      10 58.2 5.50
#> 107  8.02  7.08      10 58.2 5.50
#> 108  7.14  9.38      10 58.2 5.50
#> 109  5.68 12.10      10 58.2 5.50
#> 110  2.42 23.70      10 58.2 5.50
#> 111  0.00  0.00      11 65.0 4.92
#> 112  4.86  0.25      11 65.0 4.92
#> 113  7.24  0.50      11 65.0 4.92
#> 114  8.00  0.98      11 65.0 4.92
#> 115  6.81  1.98      11 65.0 4.92
#> 116  5.87  3.60      11 65.0 4.92
#> 117  5.22  5.02      11 65.0 4.92
#> 118  4.45  7.03      11 65.0 4.92
#> 119  3.62  9.03      11 65.0 4.92
#> 120  2.69 12.12      11 65.0 4.92
#> 121  0.86 24.08      11 65.0 4.92
#> 122  0.00  0.00      12 60.5 5.30
#> 123  1.25  0.25      12 60.5 5.30
#> 124  3.96  0.50      12 60.5 5.30
#> 125  7.82  1.00      12 60.5 5.30
#> 126  9.72  2.00      12 60.5 5.30
#> 127  9.75  3.52      12 60.5 5.30
#> 128  8.57  5.07      12 60.5 5.30
#> 129  6.59  7.07      12 60.5 5.30
#> 130  6.11  9.03      12 60.5 5.30
#> 131  4.57 12.05      12 60.5 5.30
#> 132  1.17 24.15      12 60.5 5.30