get gradients for nonmem run running via PSN

get gradients for nonmem run running via PSN

check_gradients(runnum, wd = NULL, folder_str = ".mod.dir.1",
  full_directory = NULL, print = FALSE)

Arguments

runnum

Xpose-style run number

wd

relative directly location where PSN run results are output

folder_str

folder name besides run number

full_directory

can be specified if folder where gradient file stored non-traditional

print

print visual plot of gradients

Details

by default, the working directory should be set to the location of .mod file associated with the run

the default folder structure anticipates runs were run via psn with nomenclature such as run001.mod -mod

has not been checked for side effects if have multiple runs with the same file name due to restarts or other changes

default to not print visual of gradients

Examples

## Not run: ------------------------------------
# check_gradients(001)
# check_gradients(001, wd = "~/ProjX/NONMEM")
# check_gradients(001, folder_str = "modelfit_dir1")
# check_gradients(001, full_directory = "~/user/ProjX/NONMEM/run001.mod.dir.1/NM_run2/psn.grd")
## ---------------------------------------------