chunk dataframes so easy to split for parallel processing

chunk dataframes so easy to split for parallel processing

chunk_df(.gdf, ..., .nchunks = NULL)

Arguments

.gdf

(grouped) data frame

...

grouping variables, either a character vector or NSE-style column names

.nchunks

set number of chunks

Examples

library(dplyr)
Theoph %>% chunk_df()

#> Source: local data frame [132 x 6]
#> Groups: sim_chunk [4]
#> 
#>    Subject    Wt  Dose  Time  conc sim_chunk
#>      <ord> <dbl> <dbl> <dbl> <dbl>     <int>
#> 1        1  79.6  4.02  0.00  0.74         1
#> 2        1  79.6  4.02  0.25  2.84         1
#> 3        1  79.6  4.02  0.57  6.57         1
#> 4        1  79.6  4.02  1.12 10.50         1
#> 5        1  79.6  4.02  2.02  9.66         1
#> 6        1  79.6  4.02  3.82  8.58         1
#> 7        1  79.6  4.02  5.10  8.36         1
#> 8        1  79.6  4.02  7.03  7.47         1
#> 9        1  79.6  4.02  9.05  6.89         1
#> 10       1  79.6  4.02 12.12  5.94         1
#> # ... with 122 more rows

Theoph %>% group_by(Subject) %>% chunk_df()

#> Source: local data frame [132 x 6]
#> Groups: sim_chunk [4]
#> 
#>    Subject    Wt  Dose  Time  conc sim_chunk
#>      <ord> <dbl> <dbl> <dbl> <dbl>     <int>
#> 1        1  79.6  4.02  0.00  0.74         1
#> 2        1  79.6  4.02  0.25  2.84         1
#> 3        1  79.6  4.02  0.57  6.57         1
#> 4        1  79.6  4.02  1.12 10.50         1
#> 5        1  79.6  4.02  2.02  9.66         1
#> 6        1  79.6  4.02  3.82  8.58         1
#> 7        1  79.6  4.02  5.10  8.36         1
#> 8        1  79.6  4.02  7.03  7.47         1
#> 9        1  79.6  4.02  9.05  6.89         1
#> 10       1  79.6  4.02 12.12  5.94         1
#> # ... with 122 more rows

Theoph %>% chunk_df(Subject)

#> Source: local data frame [132 x 6]
#> Groups: sim_chunk [4]
#> 
#>    Subject    Wt  Dose  Time  conc sim_chunk
#>      <ord> <dbl> <dbl> <dbl> <dbl>     <int>
#> 1        1  79.6  4.02  0.00  0.74         1
#> 2        1  79.6  4.02  0.25  2.84         1
#> 3        1  79.6  4.02  0.57  6.57         1
#> 4        1  79.6  4.02  1.12 10.50         1
#> 5        1  79.6  4.02  2.02  9.66         1
#> 6        1  79.6  4.02  3.82  8.58         1
#> 7        1  79.6  4.02  5.10  8.36         1
#> 8        1  79.6  4.02  7.03  7.47         1
#> 9        1  79.6  4.02  9.05  6.89         1
#> 10       1  79.6  4.02 12.12  5.94         1
#> # ... with 122 more rows

Theoph %>% chunk_df(c("Subject"))

#> Source: local data frame [132 x 6]
#> Groups: sim_chunk [4]
#> 
#>    Subject    Wt  Dose  Time  conc sim_chunk
#>      <ord> <dbl> <dbl> <dbl> <dbl>     <int>
#> 1        1  79.6  4.02  0.00  0.74         1
#> 2        1  79.6  4.02  0.25  2.84         1
#> 3        1  79.6  4.02  0.57  6.57         1
#> 4        1  79.6  4.02  1.12 10.50         1
#> 5        1  79.6  4.02  2.02  9.66         1
#> 6        1  79.6  4.02  3.82  8.58         1
#> 7        1  79.6  4.02  5.10  8.36         1
#> 8        1  79.6  4.02  7.03  7.47         1
#> 9        1  79.6  4.02  9.05  6.89         1
#> 10       1  79.6  4.02 12.12  5.94         1
#> # ... with 122 more rows

Theoph %>% chunk_df(Subject, .nchunks = 3)

#> Source: local data frame [132 x 6]
#> Groups: sim_chunk [3]
#> 
#>    Subject    Wt  Dose  Time  conc sim_chunk
#>      <ord> <dbl> <dbl> <dbl> <dbl>     <int>
#> 1        1  79.6  4.02  0.00  0.74         1
#> 2        1  79.6  4.02  0.25  2.84         1
#> 3        1  79.6  4.02  0.57  6.57         1
#> 4        1  79.6  4.02  1.12 10.50         1
#> 5        1  79.6  4.02  2.02  9.66         1
#> 6        1  79.6  4.02  3.82  8.58         1
#> 7        1  79.6  4.02  5.10  8.36         1
#> 8        1  79.6  4.02  7.03  7.47         1
#> 9        1  79.6  4.02  9.05  6.89         1
#> 10       1  79.6  4.02 12.12  5.94         1
#> # ... with 122 more rows