resampling

resampling

resample_df(df, key_cols, strat_cols = NULL, n = NULL,
  key_col_name = "KEY", replace = TRUE)

Arguments

df
data frame

key_cols
key columns to resample on

strat_cols
columns to maintain proportion for stratification

n
number of unique sampled keys, defaults to match dataset

key_col_name
name of outputted key column. Default to "KEY"

replace
whether to stratify with replacement

Details

This function is valuable when generating a large simulated population where you goal is to create resampled sub-populations in addition to being able to maintain certain stratifications of factors like covariate distributions

A new keyed column will be created (defaults to name 'KEY') that contains the uniquely created new samples. This allows one to easily compare against the key'd columns. Eg, if you would like to see how many times a particular individual was resampled you can check the original ID column against the number of key's associated with that ID number.

Examples

library(PKPDmisc)
library(dplyr, quiet = TRUE)

# simple example resampling by ID maintaining Gender distribution, with 10 individuals
resample_df(sd_oral_richpk, key_cols = "ID", strat_cols = "Gender", n = 10)

#> # A tibble: 120 x 10
#>      KEY    ID  Time   Amt      Conc      Age   Weight Gender   Race  Dose
#>    <int> <int> <dbl> <int>     <dbl>    <dbl>    <dbl> <fctr> <fctr> <int>
#>  1     1    26  0.00  5000   0.00000 41.26571 56.59549 Female  Black  5000
#>  2     1    26  0.25     0  27.73210 41.26571 56.59549 Female  Black  5000
#>  3     1    26  0.50     0  53.33475 41.26571 56.59549 Female  Black  5000
#>  4     1    26  1.00     0  82.48312 41.26571 56.59549 Female  Black  5000
#>  5     1    26  2.00     0 101.91661 41.26571 56.59549 Female  Black  5000
#>  6     1    26  3.00     0  81.67875 41.26571 56.59549 Female  Black  5000
#>  7     1    26  4.00     0  89.99290 41.26571 56.59549 Female  Black  5000
#>  8     1    26  6.00     0  73.01971 41.26571 56.59549 Female  Black  5000
#>  9     1    26  8.00     0  60.02791 41.26571 56.59549 Female  Black  5000
#> 10     1    26 12.00     0  48.67562 41.26571 56.59549 Female  Black  5000
#> # ... with 110 more rows

# for a more complex example lets resample "simulated" data with multiple replicates
subset_data <- sd_oral_richpk %>%
   filter(ID < 20)

# make 'simulated' data with 5 replicates and combine to single dataframe
rep_dat <- lapply(1:5, function(x) {
subset_data %>%
  mutate(REP = x)
  }) %>% bind_rows()

# now when we resample we also want to maintain the ID+REP relationship as resampling
# just the ID would give all rows associated for an ID with all reps, rather than 
# a single "unit" of ID/REP
resample_df(rep_dat, key_cols = c("ID", "REP"))

#> # A tibble: 1,140 x 11
#>      KEY    ID  Time   Amt     Conc      Age   Weight Gender      Race  Dose
#>    <int> <int> <dbl> <int>    <dbl>    <dbl>    <dbl> <fctr>    <fctr> <int>
#>  1     1     4  0.00  5000  0.00000 46.87347 62.47354 Female Caucasian  5000
#>  2     1     4  0.25     0 29.30308 46.87347 62.47354 Female Caucasian  5000
#>  3     1     4  0.50     0 50.49399 46.87347 62.47354 Female Caucasian  5000
#>  4     1     4  1.00     0 66.44664 46.87347 62.47354 Female Caucasian  5000
#>  5     1     4  2.00     0 88.99138 46.87347 62.47354 Female Caucasian  5000
#>  6     1     4  3.00     0 75.05583 46.87347 62.47354 Female Caucasian  5000
#>  7     1     4  4.00     0 96.68389 46.87347 62.47354 Female Caucasian  5000
#>  8     1     4  6.00     0 56.19676 46.87347 62.47354 Female Caucasian  5000
#>  9     1     4  8.00     0 66.60407 46.87347 62.47354 Female Caucasian  5000
#> 10     1     4 12.00     0 38.96617 46.87347 62.47354 Female Caucasian  5000
#> # ... with 1,130 more rows, and 1 more variables: REP <int>

# check to see that stratification is maintained
rep_dat %>% group_by(Gender) %>% tally

#> # A tibble: 2 x 2
#>   Gender     n
#>   <fctr> <int>
#> 1 Female   300
#> 2   Male   840

resample_df(rep_dat, key_cols=c("ID", "REP"), strat_cols="Gender") %>%
  group_by(Gender) %>% tally

#> # A tibble: 2 x 2
#>   Gender     n
#>   <fctr> <int>
#> 1 Female   300
#> 2   Male   840

rep_dat %>% group_by(Gender, Race) %>% tally

#> # A tibble: 8 x 3
#> # Groups:   Gender [?]
#>   Gender      Race     n
#>   <fctr>    <fctr> <int>
#> 1 Female Caucasian   180
#> 2 Female  Hispanic    60
#> 3 Female     Other    60
#> 4   Male     Asian   120
#> 5   Male     Black   180
#> 6   Male Caucasian    60
#> 7   Male  Hispanic   180
#> 8   Male     Other   300

resample_df(rep_dat, key_cols=c("ID", "REP"), strat_cols=c("Gender", "Race")) %>%
  group_by(Gender, Race) %>% tally

#> # A tibble: 8 x 3
#> # Groups:   Gender [?]
#>   Gender      Race     n
#>   <fctr>    <fctr> <int>
#> 1 Female Caucasian   180
#> 2 Female  Hispanic    60
#> 3 Female     Other    60
#> 4   Male     Asian   120
#> 5   Male     Black   180
#> 6   Male Caucasian    60
#> 7   Male  Hispanic   180
#> 8   Male     Other   300