cmax summary function

cmax summary function

s_cmax(df, DV, na.rm = T, name = "CMAX", carry = FALSE,
  check_duplicates = TRUE)

Arguments

df

data frame. Can pass in grouped data frame to calculate by group

DV

dependent variable column to calculate cmax on

na.rm

boolean whether to remove NA values when returning max value default = TRUE

name

string name of output column. Default to `CMAX`

carry

boolean whether to carry other columns. Defaults to FALSE

check_duplicates

boolean whether to check for duplicates in each group after calculation

Details

one trick for better performance is to turn off check duplicates the function is already quite fast by using dplyr under the hood, but if used in a simulation context where it may be used over and over, turning this off will result in slightly better performance as it does not perform the additional calculations to check for duplicate entries.

The duplicates warning is due to the nature of cmax calculations often being used in subsequent steps, either to see the time that cmax was reached, or for further calculations. Duplicate values per group (eg ID) can lead to odd behavior in later calculations, thus duplicates are checked for and the user is warned by default.

Examples

## Not run: ------------------------------------
# library(PKPDdatasets)
# pr_id <- dplyr::group_by(pain_relief, ID)
# s_cmax(pr_id, "CONC") 
# s_cmax(pr_id, "CONC", carry=T) 
# s_cmax(pr_id, "CONC", name = "max_conc") 
## ---------------------------------------------