summarize quantiles

summarize quantiles

summarize quantiles for a column

s_quantiles_(.data, x, probs, na_rm = TRUE)

s_quantiles(.data, x, probs, na_rm = TRUE)

s_quantiles_(.data, x, probs, na_rm = TRUE)

s_quantiles(.data, x, probs, na_rm = TRUE)

Arguments

.data
data frame

x
column to calculate quantiles for

probs
probability for quantiles of interest

na_rm
remove na's before calculating value for quantile

df
data frame

col_name
string name for column to calculate quantiles on

na.rm
whether to remove na values

quantiles
probabilities to calculate quantiles for

Details

can calculate multiple quantiles by passing in a vector of probabilities. Can also pass in a grouped df using dplyr::group_by to get summaries by group

Examples

library(dplyr)
sd_oral_richpk  %>% s_quantiles(Conc, probs = c(0.1, 0.5))

#>   Conc_q10 Conc_q50
#> 1 1.631003 29.88521

sd_oral_richpk  %>% group_by(Gender) %>% s_quantiles_("Conc", probs = c(0.1, 0.5))

#> # A tibble: 2 x 3
#> # Groups:   Gender [?]
#>   Gender Conc_q10 Conc_q50
#>   <fctr>    <dbl>    <dbl>
#> 1 Female 1.508296 31.25790
#> 2   Male 1.901556 28.97848

library(dplyr)
sd_oral_richpk %>% group_by(Gender, Time) %>% s_quantiles(Conc, c(0.05, 0.5, 0.95))

#> # A tibble: 24 x 5
#> # Groups:   Gender [?]
#>    Gender  Time   Conc_q5 Conc_q50  Conc_q95
#>    <fctr> <dbl>     <dbl>    <dbl>     <dbl>
#>  1 Female  0.00  0.000000  0.00000   0.00000
#>  2 Female  0.25  8.266324 19.62901  40.93315
#>  3 Female  0.50 16.474862 30.26571  62.60952
#>  4 Female  1.00 25.800352 47.54689 108.90100
#>  5 Female  2.00 36.338094 48.48951 105.16014
#>  6 Female  3.00 36.800376 50.83126  95.82737
#>  7 Female  4.00 27.461765 48.88379 106.87681
#>  8 Female  6.00 18.837319 35.08397  81.03672
#>  9 Female  8.00 16.192136 27.73593  66.71031
#> 10 Female 12.00  5.344051 24.76458  59.63218
#> # ... with 14 more rows