given a set of bin ranges, assign each value to a bin

given a set of bin ranges, assign each value to a bin

set_bins_cpp(x, left, right)

Arguments

x

A numeric vector of values

left, right

Boundary values

Details

Given a set of quantiles/bins/etc established from a separate dataset, it can be useful to assign the same bins to new or simulated data for comparisons or to do additional analysis such as assign dropouts etc. This function can be used to take the breakpoints to establish bins quickly and easily