strip additional attributes that make dplyr fail

strip additional attributes that make dplyr fail

strip_attributes(df, attr_names)

Arguments

df

dataframe

attr_names

names of attributes that you want to remove

Details

dplyr as of 0.4 still does not handle columns with non-generic attributes and will error out rather than ignoring them etc. This function will allow one to strip attribute names to allow the data frame to be used within the dplyr pipeline without issue.

This type of data is common when dealing with SAS datasets

Examples

## Not run: ------------------------------------
# foo <- data.frame(a = 1:5, b = 1:5, c=letters[1:5])
# df <- foo
# attr(df$a, "label") <- "col a"
# attr(df$b, "label") <- "col b"
# attr(df$c, "label") <- "col c"
# 
# library(dplyr)
# df %>% filter(a %in% c(1, 2)) # will throw an error
# df %>% strip_attributes("label") %>% filter(a %in% c(1, 2))
# 
# attr(df$a, "notes") <- "a note"
# # now column a has attributes of label and notes
# df %>% strip_attributes(c("label", "notes")) %>% filter(a %in% c(1, 2))
## ---------------------------------------------