Set options for annotating and saving tables and figures

mrg_script() sets the mrg.script option using this_file_proj().

tables_to() sets the pmtables.dir option as the default table output directory; this function also sets the pmtables.path.type option, defaulting to proj.

figures_to() sets the mrggsave.dir option as the default figure output directory.

tf_options() prints the mrg.script, pmtables.dir, and mrggsave.dir options to the console.

tf_options_clear() clears options set by mrg_script(), tables_to(), and figures_to(), setting the option value to NULL.

tf_options()

tf_options_clear(quietly = FALSE)

mrg_script(path = NULL)

tables_to(path, create = FALSE, set_script = TRUE, path.type = "proj")

figures_to(path, create = FALSE, set_script = TRUE)

Arguments

quietly

if TRUE, don't echo tf_options() outputs after options are cleared with tf_options_clear().

path

the absolute path or path relative to the current working directory.

create

if TRUE, the directory will be created with fs::dir_create() in case it does not already exist.

set_script

if TRUE (the default), include a call to mrg_script() to set the mrg.script option.

path.type

indicates how pmtables should format the path portion of table annotations; see pmtables::format_table_path.

Value

tables_to(), figures_to() and mrg_script() return the value of the option invisibly

tf_options() prints messages to the console and returns NULL invisibly

tf_options_clear() prints tf_options() output unless the user requests quiet reset.