Return the current script's file name or directory

The following functions use this.path to determine the file name of the current script:

this_file_name() returns the base name of the current script.

this_file_path() returns the absolute path of the current script.

this_file_proj() returns the path of the current script relative to the root of the project that contains it, as defined by this.path::this.proj().

The following functions use this.path to determine the parent directory of the current script:

this_dir_name() returns the base name of the directory containing the current script.

this_dir_path() returns the absolute path of the directory containing the current script.

this_dir_proj() returns the path of the directory containing the current script relative to the root of the project that contains it, as defined by this.path::this.proj().

this_file_name()

this_dir_name()

this_file_path()

this_dir_path()

this_file_proj()

this_dir_proj()

Value

path The file name of the current script.